gemeni_csv/requirements.txt
streamlit==1.33.0
pandas
torch
sentence-transformers
chromadb==0.4.24
langchain==0.2.5
langchain-core==0.2.9
langchain-community==0.2.4
langchain-huggingface==0.0.3
langchain-chroma>=0.1.1
huggingface-hub>=0.23.0
PyYAML